% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Crowther2003.R
\docType{data}
\name{Crowther2003}
\alias{Crowther2003}
\title{Aspirin after Myocardial Infarction}
\format{A data frame with the following columns:

\tabular{rl}{
\bold{\emph{study}}\tab study label \cr
\bold{\emph{pvh.e}}\tab number of periventricular haemorrhages in experimental group \cr
\bold{\emph{n.e}}\tab number of observations in experimental group \cr
\bold{\emph{pvh.c}}\tab number of periventricular haemorrhages in control group \cr
\bold{\emph{n.c}}\tab number of observations in control group \cr
}}
\source{
Crowther CA, Henderson-Smart DJ (2003):
Phenobarbital prior to preterm birth for preventing neonatal
periventricular haemorrhage.
\emph{Cochrane Database of Systematic Reviews}, CD000164
}
\usage{
Crowther2003
}
\description{
Meta-analysis on phenobarbital prior to preterm birth for
preventing neonatal periventricular haemorrhage
}
\examples{
data(Crowther2003)
metabin(pvh.e, n.e, pvh.c, n.c, data = Crowther2003, studlab = study)
}
\keyword{datasets}
